requirements.txt
matplotlib==3.7.1
mysql-connector-python==9.1.0
numerize==0.12
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pandas-profiling==3.2.0
streamlit==1.22.0
streamlit-apexjs==0.0.3
streamlit-card==0.0.4
streamlit-elements==0.1.0
streamlit-option-menu==0.3.2
streamlit-pandas-profiling==0.1.3
plotly==5.14.1
plotly-express==0.4.1
scikit-learn==1.2.2
seaborn==0.12.2
pandas-profiling==3.2.0
seaborn==0.12.2
streamlit-extras==0.2.7
streamlit-card==0.0.4
decorator==5.1.1
deprecation==2.1.0
numerize==0.12
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
